\name{isoTensor-package}
\alias{isoTensor-package}
\docType{package}

\title{
\packageTitle{isoTensor}
}
\description{
\packageDescription{isoTensor}
}
\details{
The main function is \code{\link{isoTensor}}, which estimates the
population-level cell-type-resolved isoform abundance matrix B from a
bulk isoform abundance matrix X and a fixed, externally estimated cell
type fraction matrix A. Toy data is available from
\code{\link{isoToyModel}}.
}
\author{
\packageAuthor{isoTensor}

Maintainer: \packageMaintainer{isoTensor}
}

\keyword{package}
\seealso{\code{\link{isoTensor}}, \code{\link{isoToyModel}}}
